requirements.txt
absl-py==2.3.1
astunparse==1.6.3
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
mysql-connector-python==9.0.0
Flask-WTF==1.2.1
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.14.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.10.0
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
mysql-connector-python==9.0.0
namex==0.1.0
numpy==2.3.5
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
protobuf==6.33.0
Pygments==2.19.2
python-dotenv==1.0.1
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
midtransclient==1.4.2